refactor(agentic): extract verification_scraper and streamline parameter signatures (#93, #94)
